requirements.txt
nodriver
pandas
yagmail
python-dotenv
openai